+++ services/web/server/src/simcore_service_webserver/meta_modeling_version_control.py
@@ -11,13 +11,13 @@ from models_library.utils.fastapi_encoders import jsonable_encoder
 
 from .projects.project_models import ProjectDict
 from .version_control.errors import UserUndefined
+from .version_control.models import CommitID, TagProxy
 from .version_control.vc_changes import (
     compute_workbench_checksum,
     eval_workcopy_project_id,
 )
 from .version_control.vc_tags import compose_workcopy_project_tag_name
 from .version_control.version_control_db import VersionControlRepository
-from .version_control.version_control_models import CommitID, TagProxy
 
 log = logging.getLogger(__name__)
