+++ services/web/server/src/simcore_service_webserver/version_control/_core.py
@@ -15,8 +15,8 @@ from aiopg.sa.result import RowProxy
 from pydantic import NonNegativeInt, PositiveInt, validate_arguments
 
 from .errors import CleanRequiredError
+from .models import Checkpoint, CommitLog, RefID, WorkbenchView
 from .version_control_db import VersionControlRepository
-from .version_control_models import Checkpoint, CommitLog, RefID, WorkbenchView
 
 log = logging.getLogger(__name__)
