+++ services/web/server/tests/unit/with_dbs/03/version_control/test_version_control_handlers.py
@@ -15,7 +15,7 @@ from models_library.users import UserID
 from pydantic.main import BaseModel
 from pytest_simcore.helpers.utils_assert import assert_status
 from simcore_service_webserver._meta import API_VTAG as VX
-from simcore_service_webserver.version_control.version_control_models import (
+from simcore_service_webserver.version_control.models import (
     CheckpointApiModel,
     RepoApiModel,
 )
